pyproject.toml
[project]
name = "korrekturabzug"
version = "1.0.0"
description = "PDF Preflight Analysis & Proof Sheet Generator"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.100,<1",
    "uvicorn[standard]>=0.20,<1",
    "python-multipart>=0.0.6,<1",
    "pikepdf>=8,<10",
    "PyMuPDF>=1.23,<2",
    "reportlab>=4,<5",
    "pydantic>=2,<3",
    "pyyaml>=6,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7,<9",
    "httpx>=0.24,<1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.backends._legacy:_Backend"
